Keep the server stdin open in docker
It appears that part of the react stack (I saw multiple references to
react-scripts in stackoverflow) assumes it will run in a TTY and as such
will terminate if STDIN is closed. Docker (Compose?) does not keep STDIN
open by default (for a service specifically?) so direct the compose
manifest to do so.